SelectDropdown: instanceId doesn't need to be in state and be used for keys
Changes `instanceId` to be an instance property assigned during construction. No need
to put it in state, as it never changes.

Also, don't use `instanceId` to create `key` values. These don't need to be globally unique.
Use them only for DOM element IDs that are used to link elements together with ARIA attributes.