Start replacing 'url' and 'qs' with standard APIs.
We currently make use of node's deprecated `'url'` API in Calypso
client, imported as a browserified module.
We also make use of a 3rd party `'qs'` library to parse and format
URL query strings.

This PR starts the process of replacing them both (in client code) with
native functionality provided by `URL` and `URLSearchParams`.

It uses a polyfill for browsers that don't include native support.
Modern browsers don't need this polyfill, so it's excluded from the
evergreen build.

In addition, the PR tweaks `lib/build-url` to use a more semantic name
for its single export, preparing it also for future extension.