FSE: Decode site option before rendering
Since wordpress encoded these elements on the backend,
they do not appear to be decoded by the API fetch. As a
result, we have to decode the string ourselves before using it.

Thankfully, though, we do not have to encode it, since that is
already handled at some point in the backend.